+++ scripts/logtalk_tester.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   Unit testing automation script
-##   Last updated on March 16, 2016
+##   Last updated on March 24, 2016
 ## 
 ##   This file is part of Logtalk <http://logtalk.org/>  
 ##   Copyright 1998-2016 Paulo Moura <pmoura@logtalk.org>
@@ -25,7 +25,7 @@
 # loosely based on a unit test automation script contributed by Parker Jones
 
 print_version() {
-	echo "$(basename "$0") 0.6"
+	echo "$(basename "$0") 0.7"
 	exit 0
 }
 
@@ -84,6 +84,7 @@ format='default'
 format_goal=$format_default_goal
 # disable timeouts to maintain backward compatibility
 timeout=0
+arguments=""
 
 run_tests() {
 	unit=$(dirname "$1")
@@ -121,9 +122,9 @@ run_test() {
 	name="$1"
 	goal="$2"
 	if [ "$timeout_command" != "" ] && [ $timeout -ne 0 ] ; then
-		$timeout_command $timeout $logtalk_call "$goal" > "$results/$name.results" 2> "$results/$name.errors"
+		$timeout_command $timeout $logtalk_call "$goal" -- "$arguments" > "$results/$name.results" 2> "$results/$name.errors"
 	else
-		$logtalk_call "$goal" > "$results/$name.results" 2> "$results/$name.errors"
+		$logtalk_call "$goal" -- "$arguments" > "$results/$name.results" 2> "$results/$name.errors"
 	fi
 }
 
@@ -135,7 +136,7 @@ usage_help()
 	echo  "case of failed unit tests, this script returns an exit code of 1."
 	echo
 	echo "Usage:"
-	echo "  $(basename "$0") [-p prolog] [-m mode] [-f format] [-d results] [-t timeout]"
+	echo "  $(basename "$0") [-p prolog] [-m mode] [-f format] [-d results] [-t timeout] [-a arguments]"
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -h"
 	echo
@@ -149,12 +150,13 @@ usage_help()
 	echo "     (possible values are default, tap, and xunit)"
 	echo "  -d directory to store the test logs (default is ./logtalk_tester_logs)"
 	echo "  -t timeout in seconds for running each test set (default is $timeout; i.e. disabled)"
+	echo "  -a arguments to be passed to the integration script used to run the tests (no default)"
 	echo "  -h help"
 	echo
 	exit 0
 }
 
-while getopts "vp:m:f:d:t:h" option
+while getopts "vp:m:f:d:t:a:h" option
 do
 	case $option in
 		v) print_version;;
@@ -163,6 +165,7 @@ do
 		f) f_arg="$OPTARG";;
 		d) d_arg="$OPTARG";;
 		t) t_arg="$OPTARG";;
+		a) a_arg="$OPTARG";;
 		h) usage_help;;
 		*) usage_help;;
 	esac
@@ -281,6 +284,10 @@ fi
 
 if [ "$t_arg" != "" ] ; then
 	timeout="$t_arg"
+fi
+
+if [ "$a_arg" != "" ] ; then
+	arguments="$a_arg"
 fi
 
 if [ "$timeout_command" == "" ] ; then